Refactoring Server and separating concerns
